Changed datatype to uint16

+++ NetworkManagerContract.sol
@@ -3,14 +3,13 @@ contract NetworkManagerContract {
 
     uint nodeCounter;
   
-    struct NodeDetails {
-  
+    struct NodeDetails {  
         string nodeName;
         string role; 
         string publickey;
         string enode;
-	string ip;
-	string id;
+        string ip;
+        string id;
     }
 
     mapping (string => NodeDetails)nodes;
@@ -23,23 +22,23 @@ contract NetworkManagerContract {
         nodes[e].publickey = p;
         nodes[e].nodeName = n;
         nodes[e].role = r;
-	nodes[e].ip = ip;
-	nodes[e].id = id;
+        nodes[e].ip = ip;
+        nodes[e].id = id;
         enodeList.push(e);
-        emit print(n,r,p,e,ip,id);
+        emit print(n, r, p, e, ip, id);
     
     }
    
-    function getNodeDetails(uint _index) constant public returns (string n, string r,string p,string ip,string e,string id,uint i) {
+    function getNodeDetails(uint16 _index) constant public returns (string n, string r,string p,string ip,string e,string id,uint i) {
         NodeDetails memory nodeInfo = nodes[enodeList[_index]];
         return (
-                nodeInfo.nodeName,
-                nodeInfo.role,
-                nodeInfo.publickey,
-	        nodeInfo.ip,
-                enodeList[_index],
-		nodeInfo.id,
-                _index
+            nodeInfo.nodeName,
+            nodeInfo.role,
+            nodeInfo.publickey,
+            nodeInfo.ip,
+            enodeList[_index],
+            nodeInfo.id,
+            _index
         );
     }
 
@@ -52,23 +51,22 @@ contract NetworkManagerContract {
         nodes[e].publickey = p;
         nodes[e].nodeName = n;
         nodes[e].role = r;
-	nodes[e].ip = ip;
-	nodes[e].id = id;
-        emit print(n,r,p,e,ip,id);
+        nodes[e].ip = ip;
+        nodes[e].id = id;
+        emit print(n, r, p, e, ip, id);
     }
 
-    function getNodeList(uint i)  public  constant   returns (string n, string r,string p,string ip,string e, string id) {
+    function getNodeList(uint16 i)  public  constant   returns (string n, string r,string p,string ip,string e, string id) {
         
         NodeDetails memory nodeInfo = nodes[enodeList[i]];
         return (
             nodeInfo.nodeName,
             nodeInfo.role,
             nodeInfo.publickey,
-	    nodeInfo.ip,
+	        nodeInfo.ip,
             enodeList[i],
             nodeInfo.id
         );
-        
     }
 
 }
